+++ lib/src/animated_stack_manager.dart
@@ -8,11 +8,17 @@ import '../animated_stack_widget.dart';
 /// the animated list that belongs to [animatedStackKey].
 ///
 class AnimatedStackManager<E> {
+  static const Duration _kDuration = Duration(milliseconds: 300);
+
+  final Duration duration;
+
   AnimatedStackManager({
     required this.animatedStackKey,
     required this.removedItemBuilder,
     Iterable<E>? initialItems,
-  }) : _items = List<E>.from(initialItems ?? <E>[]);
+    Duration? duration,
+  })  : _items = List<E>.from(initialItems ?? <E>[]),
+        duration = duration ?? _kDuration;
 
   final GlobalKey<AnimatedStackState> animatedStackKey;
   final RemovedItemBuilder<E> removedItemBuilder;
@@ -20,12 +26,12 @@ class AnimatedStackManager<E> {
 
   AnimatedStackState? get _animatedStack => animatedStackKey.currentState;
 
-  void insert(int index, E item) {
+  void insert(int index, E item, {Duration? duration}) {
     _items.insert(index, item);
-    _animatedStack!.insertItem(index);
+    _animatedStack!.insertItem(index, duration: duration ?? this.duration);
   }
 
-  E removeAt(int index) {
+  E removeAt(int index, {Duration? duration}) {
     final E removedItem = _items.removeAt(index);
     if (removedItem != null) {
       _animatedStack!.removeItem(
@@ -33,12 +39,13 @@ class AnimatedStackManager<E> {
         (BuildContext context, Animation<double> animation) {
           return removedItemBuilder(removedItem, context, animation);
         },
+        duration: duration ?? this.duration,
       );
     }
     return removedItem;
   }
 
-  void clear() {
+  void clear({Duration? duration}) {
     for (var i = 0; i <= _items.length - 1; i++) {
       final item = _items[i];
       _animatedStack!.removeItem(
@@ -46,6 +53,7 @@ class AnimatedStackManager<E> {
         (BuildContext context, Animation<double> animation) {
           return removedItemBuilder(item, context, animation);
         },
+        duration: duration ?? this.duration,
       );
     }
     _items.clear();
@@ -56,6 +64,12 @@ class AnimatedStackManager<E> {
   E operator [](int index) => _items[index];
 
   int indexOf(E item) => _items.indexOf(item);
+
+  bool any(bool Function(E) test) => _items.any(test);
+
+  bool get isEmpty => _items.isEmpty;
+
+  bool get isNotEmpty => _items.isNotEmpty;
 }
 
 typedef RemovedItemBuilder<E> = Widget Function(